Upload of original tested site
At this upload, the video div had been commented out on the training page and needs to be corrected or worked on.  I was currently trying to add a carousel to the tacticstraining.html file for testing purposes and once working will copy it to the other pages. So the tacticstraining page is slightly broken on this initial upload